+++ app/lib/utils/serviceDataMapper.js
@@ -1,8 +1,20 @@
+function joinTruthyValues(obj) {
+  return Object.values(obj)
+    .filter(value => value)
+    .join(', ');
+}
+
 function addressFormatter(address) {
-  if ((address) && (address.addressLines) && (address.postcode)) {
-    return `${address.addressLines.join(', ')}, ${address.postcode}`;
+  if (!address || Object.keys(address).length === 0) { return undefined; }
+
+  let addressLines = [];
+  if (address.addressLines) {
+    addressLines = address.addressLines;
   }
-  return undefined;
+  if (address.postcode) {
+    addressLines.push(address.postcode);
+  }
+  return joinTruthyValues(addressLines);
 }
 
 module.exports = {