+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - Replace sexperts search with kit locations for over 25s
 - Use `arc` rather than `plane` for precision in distance
 - Use location input for map link
-- Use name prior to special characters in map link
+- Do not use name of place in map link
+- Display address when there is no postcode
 
 0.5.0 / 2018-04-11
 ==================